+++ util/CloudUtil.go
@@ -1,7 +1,7 @@
 package util
 
 import (
-	"github.com/devtron-labs/ci-runner/helper"
+	"github.com/caarlos0/env"
 	blob_storage "github.com/devtron-labs/common-lib/blob-storage"
 )
 
@@ -10,6 +10,47 @@ const (
 	BlobStorageObjectTypeArtifact = "artifact"
 	BlobStorageObjectTypeLog      = "log"
 )
+
+// BlobStorageConfig is the blob storage config for external cluster added via cm/secret code
+// will be expecting these env variables acc to cloud provider if UseExternalClusterBlob is true.
+type BlobStorageConfig struct {
+	//AWS credentials
+	CloudProvider      blob_storage.BlobStorageType `env:"BLOB_STORAGE_PROVIDER"`
+	S3AccessKey        string                       `env:"BLOB_STORAGE_S3_ACCESS_KEY"`
+	S3SecretKey        string                       `env:"BLOB_STORAGE_S3_SECRET_KEY"`
+	S3Endpoint         string                       `env:"BLOB_STORAGE_S3_ENDPOINT"`
+	S3EndpointInsecure bool                         `env:"BLOB_STORAGE_S3_ENDPOINT_INSECURE" envDefault:"false"`
+	S3BucketVersioned  bool                         `env:"BLOB_STORAGE_S3_BUCKET_VERSIONED" envDefault:"true"`
+	//artifact and logs bucket name in s3 will get their values from DEFAULT_BUILD_LOGS_BUCKET
+	CdDefaultBuildLogsBucket string `env:"DEFAULT_BUILD_LOGS_BUCKET" `
+	//logs and artifact region in s3 will get their values from DEFAULT_CD_LOGS_BUCKET_REGION
+	CdDefaultCdLogsBucketRegion string `env:"DEFAULT_CD_LOGS_BUCKET_REGION" `
+	//cache bucket name in s3 will get its value from DEFAULT_CACHE_BUCKET
+	DefaultCacheBucket string `env:"DEFAULT_CACHE_BUCKET"`
+	//cache region in s3 will get its value from DEFAULT_CACHE_BUCKET_REGION
+	DefaultCacheBucketRegion string `env:"DEFAULT_CACHE_BUCKET_REGION"`
+
+	//GCP credentials
+	GcpBlobStorageCredentialJson string `env:"BLOB_STORAGE_GCP_CREDENTIALS_JSON"`
+	//ArtifactBucketName and LogBucketName for gcp will get their values from DEFAULT_BUILD_LOGS_BUCKET
+	//CacheBucketName for gcp will get its value from DEFAULT_CACHE_BUCKET
+
+	//Azure credentials
+	AzureAccountName               string `env:"AZURE_ACCOUNT_NAME"`
+	AzureGatewayUrl                string `env:"AZURE_GATEWAY_URL"`
+	AzureGatewayConnectionInsecure bool   `env:"AZURE_GATEWAY_CONNECTION_INSECURE" envDefault:"true"`
+	AzureAccountKey                string `env:"AZURE_ACCOUNT_KEY"`
+	//log and artifact container name in azure will get their values from AZURE_BLOB_CONTAINER_CI_LOG
+	AzureBlobContainerCiLog string `env:"AZURE_BLOB_CONTAINER_CI_LOG"`
+	//cache container name in azure will get their values from AZURE_BLOB_CONTAINER_CI_CACHE
+	AzureBlobContainerCiCache string `env:"AZURE_BLOB_CONTAINER_CI_CACHE"`
+}
+
+func GetBlobStorageConfig() (*BlobStorageConfig, error) {
+	cfg := &BlobStorageConfig{}
+	err := env.Parse(cfg)
+	return cfg, err
+}
 
 type CloudHelperBaseConfig struct {
 	StorageModuleConfigured bool
@@ -22,7 +63,7 @@ type CloudHelperBaseConfig struct {
 	BlobStorageObjectType   string
 }
 
-func (c *CloudHelperBaseConfig) SetAwsBlobStorageS3Config(blobStorageConfig *helper.BlobStorageConfig) {
+func (c *CloudHelperBaseConfig) SetAwsBlobStorageS3Config(blobStorageConfig *BlobStorageConfig) {
 	c.BlobStorageS3Config = &blob_storage.BlobStorageS3Config{
 		AccessKey:                  blobStorageConfig.S3AccessKey,
 		Passkey:                    blobStorageConfig.S3SecretKey,
@@ -40,7 +81,7 @@ func (c *CloudHelperBaseConfig) SetAwsBlobStorageS3Config(blobStorageConfig *hel
 	}
 }
 
-func (c *CloudHelperBaseConfig) SetAzureBlobStorageConfig(blobStorageConfig *helper.BlobStorageConfig) {
+func (c *CloudHelperBaseConfig) SetAzureBlobStorageConfig(blobStorageConfig *BlobStorageConfig) {
 	c.AzureBlobConfig = &blob_storage.AzureBlobConfig{
 		Enabled:               blobStorageConfig.CloudProvider == blob_storage.BLOB_STORAGE_AZURE,
 		AccountName:           blobStorageConfig.AzureAccountName,
@@ -51,7 +92,7 @@ func (c *CloudHelperBaseConfig) SetAzureBlobStorageConfig(blobStorageConfig *hel
 	}
 }
 
-func (c *CloudHelperBaseConfig) SetGcpBlobStorageConfig(blobStorageConfig *helper.BlobStorageConfig) {
+func (c *CloudHelperBaseConfig) SetGcpBlobStorageConfig(blobStorageConfig *BlobStorageConfig) {
 	c.GcpBlobConfig = &blob_storage.GcpBlobConfig{
 		CredentialFileJsonData: blobStorageConfig.GcpBlobStorageCredentialJson,
 		CacheBucketName:        blobStorageConfig.DefaultCacheBucket,
